requirements.txt
asgiref==3.10.0
Django==4.2.26
django-filter==25.1
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
